Moved ViewProfileModal to modals
